Changes to be committed: 	modified:   app/whatsapp/app.py
This code defines a configuration for a chat agent.
First, it checks if there is an environment variable called "CHAT_START_TEMPLATE_TEXT". If there is, it logs a warning message that the variable was set and assigns the value to a variable called "start_template".
If the "CHAT_START_TEMPLATE_TEXT" environment variable is not set, the code checks if there is an environment variable called "CHAT_START_TEMPLATE_FILE". If there is, it checks if a file exists at the path specified by the environment variable, and if it does, it reads the contents of the file and assigns them to the "start_template" variable. If the file does not exist, it logs a warning message indicating that it could not find the start template file.
If neither environment variable is set or if the file specified by "CHAT_START_TEMPLATE_FILE" does not exist, the code does not assign a value to "start_template".
Finally, the code logs a warning message indicating the value of "start_template".